Add Rust version
Port the Crystal version to Rust, reasonably close.
The algorithm was not changed. Progress printing has been
added for debug purposes and the macro at the beginning is a
leftover of trying to make V3 as generic as possible.

That would have required being generic over sqrt, which
would need either the num crate or another trait, and was
therefore not a good option for this port.